habits are now fully based on the database
- Toggling checkbox adds/removes entries from HabitEntry table
- Learned about and Implemented derived queries in Spring Data JPA
- Improved logging for habit status changes

+++ src/main/java/com/ozzo/habit_tracker/controller/HabitController.java
@@ -21,7 +21,7 @@ public class HabitController {
     @PostMapping("/toggle-day")
     @ResponseBody
     public ResponseEntity<Void> toggleHabitDay(
-            @RequestParam Integer habitId,
+            @RequestParam Long habitId,
             @RequestParam Boolean completed) {
 
         LocalDate today = LocalDate.now();
@@ -36,7 +36,7 @@ public class HabitController {
     @PostMapping("/toggle-week")
     @ResponseBody
     public ResponseEntity<Void> toggleHabitWeek(
-            @RequestParam Integer habitId,
+            @RequestParam Long habitId,
             @RequestParam("date") String dateStr,
             @RequestParam Boolean completed) {
